safari-534.53-branch/LayoutTests/platform/mac/fast/loader/file-url-mimetypes-2-expected.txt

example-js.js has MIME type application/x-javascript
example-jscript.jscript has MIME type text/javascript
example-kar.kar has MIME type audio/midi
example-l.l has MIME type text/plain
example-latex.latex has MIME type application/x-latex
example-lha.lha has MIME type application/octet-stream
example-lid.lid has MIME type text/plain
example-lm.lm has MIME type text/plain
example-lmm.lmm has MIME type text/plain
example-lpp.lpp has MIME type text/plain
example-lxx.lxx has MIME type text/plain
example-lzh.lzh has MIME type application/octet-stream
example-m.m has MIME type text/plain
example-m15.m15 has MIME type video/mpeg
example-m2v.m2v has MIME type video/mpeg2
example-m3u.m3u has MIME type audio/x-mpegurl
example-m4a.m4a has MIME type audio/mp4
example-m4b.m4b has MIME type application/octet-stream
example-m4p.m4p has MIME type audio/x-m4p
example-m4v.m4v has MIME type application/octet-stream
example-m75.m75 has MIME type video/mpeg
example-mac.mac has MIME type image/x-macpaint
example-man.man has MIME type application/x-troff-man
example-me.me has MIME type application/x-troff-me
example-mesh.mesh has MIME type model/mesh
example-mid.mid has MIME type audio/midi
example-midi.midi has MIME type audio/midi
example-mif.mif has MIME type application/vnd.mif
example-mig.mig has MIME type text/plain
example-mm.mm has MIME type text/plain
example-mos.mos has MIME type application/octet-stream
example-mov.mov has MIME type video/quicktime
example-movie.movie has MIME type video/x-sgi-movie
example-mp2.mp2 has MIME type audio/mpeg
example-mp3.mp3 has MIME type audio/mpeg
example-mp4.mp4 has MIME type video/mp4
example-mpe.mpe has MIME type video/mpeg
example-mpeg.mpeg has MIME type video/mpeg
example-mpg.mpg has MIME type video/mpeg
example-mpga.mpga has MIME type audio/mpeg
example-mrw.mrw has MIME type application/octet-stream
example-ms.ms has MIME type application/x-troff-ms
example-msh.msh has MIME type model/mesh
example-mxu.mxu has MIME type video/vnd.mpegurl
example-nc.nc has MIME type application/x-netcdf
example-nef.nef has MIME type application/octet-stream
example-o.o has MIME type application/octet-stream
example-oda.oda has MIME type application/oda
example-odc.odc has MIME type application/vnd.oasis.opendocument.chart
example-odf.odf has MIME type application/vnd.oasis.opendocument.formula
example-odg.odg has MIME type application/vnd.oasis.opendocument.graphics
example-odi.odi has MIME type application/vnd.oasis.opendocument.image
example-odm.odm has MIME type application/vnd.oasis.opendocument.text-master
example-odp.odp has MIME type application/vnd.oasis.opendocument.presentation
example-ods.ods has MIME type application/vnd.oasis.opendocument.spreadsheet
example-odt.odt has MIME type application/vnd.oasis.opendocument.text
example-omf.omf has MIME type application/octet-stream
example-orf.orf has MIME type application/octet-stream
example-otc.otc has MIME type application/vnd.oasis.opendocument.chart-template
example-otf.otf has MIME type application/vnd.oasis.opendocument.formula-template
example-otg.otg has MIME type application/vnd.oasis.opendocument.graphics-template
example-oth.oth has MIME type application/vnd.oasis.opendocument.text-web
example-oti.oti has MIME type application/vnd.oasis.opendocument.image-template
example-otp.otp has MIME type application/vnd.oasis.opendocument.presentation-template
example-ots.ots has MIME type application/vnd.oasis.opendocument.spreadsheet-template
example-ott.ott has MIME type application/vnd.oasis.opendocument.text-template
example-pas.pas has MIME type text/plain
example-pbm.pbm has MIME type image/x-portable-bitmap
example-pch.pch has MIME type text/plain
example-pch++.pch++ has MIME type text/plain
example-pct.pct has MIME type image/pict
example-pcx.pcx has MIME type image/x-pcx
example-pdb.pdb has MIME type chemical/x-pdb
example-pdf.pdf has MIME type application/pdf
example-pef.pef has MIME type application/octet-stream
example-pf.pf has MIME type application/octet-stream
example-pfa.pfa has MIME type application/octet-stream
example-pfb.pfb has MIME type application/octet-stream
example-pgm.pgm has MIME type image/x-portable-graymap
example-pgn.pgn has MIME type application/x-chess-pgn
example-ph3.ph3 has MIME type text/php
example-ph4.ph4 has MIME type text/php
example-php.php has MIME type text/php
example-php3.php3 has MIME type text/php
example-php4.php4 has MIME type text/php
example-phtml.phtml has MIME type text/php
example-pic.pic has MIME type image/pict
example-pict.pict has MIME type image/pict
example-pl.pl has MIME type text/x-perl-script
example-plist.plist has MIME type application/octet-stream
example-pls.pls has MIME type audio/x-scpls
example-pm.pm has MIME type text/x-perl-script
example-png.png has MIME type image/png
example-pnm.pnm has MIME type image/x-portable-anymap
example-pnt.pnt has MIME type image/x-macpaint
example-pntg.pntg has MIME type image/x-macpaint
example-potm.potm has MIME type application/vnd.ms-powerpoint.template.macroenabled.12
example-potx.potx has MIME type application/vnd.openxmlformats-officedocument.presentationml.template
example-ppm.ppm has MIME type image/x-portable-pixmap
example-ppsm.ppsm has MIME type application/vnd.ms-powerpoint.slideshow.macroenabled.12
example-ppsx.ppsx has MIME type application/vnd.openxmlformats-officedocument.presentationml.slideshow
example-ppt.ppt has MIME type application/vnd.ms-powerpoint
example-pptm.pptm has MIME type application/vnd.ms-powerpoint.presentation.macroenabled.12
example-pptx.pptx has MIME type application/vnd.openxmlformats-officedocument.presentationml.presentation
example-psd.psd has MIME type image/vnd.adobe.photoshop
example-py.py has MIME type text/x-python-script
example-qif.qif has MIME type text/qif
example-qt.qt has MIME type video/quicktime
example-qti.qti has MIME type image/x-quicktime
example-qtif.qtif has MIME type image/x-quicktime
example-qtz.qtz has MIME type application/x-quartzcomposer
example-r.r has MIME type text/plain
example-ra.ra has MIME type audio/vnd.rn-realaudio
example-raf.raf has MIME type application/octet-stream
example-ram.ram has MIME type audio/vnd.rn-realaudio
example-ras.ras has MIME type image/x-cmu-raster
example-raw.raw has MIME type application/octet-stream
example-rb.rb has MIME type text/x-ruby-script
example-rbw.rbw has MIME type text/x-ruby-script
example-rgb.rgb has MIME type image/x-rgb
example-rm.rm has MIME type application/vnd.rn-realmedia
example-roff.roff has MIME type application/x-troff
example-rpm.rpm has MIME type audio/x-pn-realaudio-plugin
example-rtf.rtf has MIME type text/rtf
example-rtx.rtx has MIME type text/richtext
Using an image element to load resources, this test makes sure that every MIME type the loader should know about based solely on file extension in file:// urls is correct.
